[refac] move SuccessCode.java and ErrorCode.java
