+++ src/main/java/org/hankki/hankkiserver/domain/user/model/Platform.java
@@ -2,7 +2,7 @@ package org.hankki.hankkiserver.domain.user.model;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.hankki.hankkiserver.common.code.ErrorCode;
+import org.hankki.hankkiserver.common.response.code.ErrorCode;
 import org.hankki.hankkiserver.common.exception.InvalidValueException;
 
 import java.util.Arrays;
